requirements.txt
appdirs==1.4.4
colorama==0.4.4
importlib-metadata==4.8.1
pyee==8.2.2
pyppeteer==0.2.6
python-dotenv==0.19.0
tqdm==4.62.2
urllib3==1.26.6
websockets==9.1
zipp==3.5.0
